Revert "[ISSUE-415][REFACTOR] Split StorageManager to different scala file"
This reverts commit 5a9e1d42dff27d02fb61708f9a2a21e941380622.
