+++ server-worker/src/main/java/com/aliyun/emr/rss/service/deploy/worker/FileWriter.java
@@ -27,7 +27,6 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
 
-import com.aliyun.emr.rss.service.deploy.worker.storage.*;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.CompositeByteBuf;
 import org.slf4j.Logger;
