+++ server-worker/src/main/scala/com/aliyun/emr/rss/service/deploy/worker/PushDataHandler.scala
@@ -23,6 +23,7 @@ import java.util.concurrent.atomic.AtomicBoolean
 
 import com.google.common.base.Throwables
 import io.netty.buffer.ByteBuf
+
 import com.aliyun.emr.rss.common.RssConf
 import com.aliyun.emr.rss.common.exception.AlreadyClosedException
 import com.aliyun.emr.rss.common.internal.Logging
@@ -34,7 +35,6 @@ import com.aliyun.emr.rss.common.network.server.BaseMessageHandler
 import com.aliyun.emr.rss.common.protocol.{PartitionLocation, PartitionSplitMode}
 import com.aliyun.emr.rss.common.protocol.message.StatusCode
 import com.aliyun.emr.rss.common.unsafe.Platform
-import com.aliyun.emr.rss.service.deploy.worker.storage.local.LocalFlusher
 
 
 class PushDataHandler extends BaseMessageHandler with Logging {
