+++ server-worker/src/main/scala/com/aliyun/emr/rss/service/deploy/worker/StorageManager.scala
@@ -15,33 +15,232 @@
  * limitations under the License.
  */
 
-package com.aliyun.emr.rss.service.deploy.worker.storage
+package com.aliyun.emr.rss.service.deploy.worker
 
 import java.io.{File, IOException}
+import java.nio.channels.{ClosedByInterruptException, FileChannel}
 import java.nio.charset.StandardCharsets
 import java.nio.file.{Files, Paths}
 import java.util
-import java.util.concurrent.atomic.AtomicInteger
-import java.util.concurrent.{ConcurrentHashMap, Executors, ThreadPoolExecutor, TimeUnit}
+import java.util.concurrent.{ConcurrentHashMap, Executors, LinkedBlockingQueue, ThreadPoolExecutor, TimeUnit}
+import java.util.concurrent.atomic.{AtomicBoolean, AtomicInteger, AtomicLongArray, LongAdder}
 import java.util.function.IntUnaryOperator
 
 import scala.collection.JavaConverters._
+import scala.util.Random
 
 import com.google.common.util.concurrent.ThreadFactoryBuilder
+import io.netty.buffer.{CompositeByteBuf, Unpooled}
 import org.iq80.leveldb.DB
+
 import com.aliyun.emr.rss.common.RssConf
 import com.aliyun.emr.rss.common.exception.RssException
 import com.aliyun.emr.rss.common.internal.Logging
-import com.aliyun.emr.rss.common.meta.{DeviceInfo, DiskInfo, DiskStatus}
+import com.aliyun.emr.rss.common.meta.{DeviceInfo, DiskInfo}
+import com.aliyun.emr.rss.common.meta.DiskStatus
 import com.aliyun.emr.rss.common.metrics.source.AbstractSource
-import com.aliyun.emr.rss.common.network.server.FileInfo
+import com.aliyun.emr.rss.common.network.server.{FileInfo, MemoryTracker}
 import com.aliyun.emr.rss.common.network.server.MemoryTracker.MemoryTrackerListener
-import com.aliyun.emr.rss.common.protocol.{PartitionLocation, PartitionSplitMode, PartitionType}
+import com.aliyun.emr.rss.common.protocol.{PartitionLocation, PartitionSplitMode, PartitionType, StorageInfo}
 import com.aliyun.emr.rss.common.util.{ThreadUtils, Utils}
 import com.aliyun.emr.rss.common.utils.PBSerDeUtils
-import com.aliyun.emr.rss.service.deploy.worker._
-import com.aliyun.emr.rss.service.deploy.worker.storage.local.LocalFlusher
+import com.aliyun.emr.rss.service.deploy.worker.FileWriter.FlushNotifier
 
+trait DeviceObserver {
+  def notifyError(mountPoint: String, diskStatus: DiskStatus): Unit = {}
+  def notifyHealthy(mountPoint: String): Unit = {}
+  def notifyHighDiskUsage(mountPoint: String): Unit = {}
+}
+
+private[worker] abstract class FlushTask(
+    val buffer: CompositeByteBuf,
+    val notifier: FlushNotifier) {
+  def flush(): Unit
+}
+
+private[worker] class LocalFlushTask(
+    buffer: CompositeByteBuf,
+    fileChannel: FileChannel,
+    notifier: FileWriter.FlushNotifier) extends FlushTask(buffer, notifier) {
+  override def flush(): Unit = {
+    fileChannel.write(buffer.nioBuffers())
+  }
+}
+
+private[worker] abstract class Flusher(
+    val workerSource: AbstractSource,
+    val threadCount: Int,
+    val flushAvgTimeWindowSize: Int,
+    val flushAvgTimeMinimumCount: Int) extends Logging {
+  protected lazy val flusherId = System.identityHashCode(this)
+  protected val workingQueues = new Array[LinkedBlockingQueue[FlushTask]](threadCount)
+  protected val bufferQueue = new LinkedBlockingQueue[CompositeByteBuf]()
+  protected val workers = new Array[Thread](threadCount)
+  protected var nextWorkerIndex: Int = 0
+  protected val flushCount = new LongAdder
+  protected val flushTotalTime = new LongAdder
+  protected val avgTimeWindow = new Array[(Long, Long)](flushAvgTimeWindowSize)
+  protected var avgTimeWindowCurrentIndex = 0
+
+  val lastBeginFlushTime: AtomicLongArray = new AtomicLongArray(threadCount)
+  val stopFlag = new AtomicBoolean(false)
+  val rand = new Random()
+
+  init()
+
+  private def init(): Unit = {
+    for (i <- 0 until flushAvgTimeWindowSize) {
+      avgTimeWindow(i) = (0L, 0L)
+    }
+    for (i <- 0 until lastBeginFlushTime.length()) {
+      lastBeginFlushTime.set(i, -1)
+    }
+    for (index <- 0 until threadCount) {
+      workingQueues(index) = new LinkedBlockingQueue[FlushTask]()
+      workers(index) = new Thread(s"$this-$index") {
+        override def run(): Unit = {
+          while (!stopFlag.get()) {
+            val task = workingQueues(index).take()
+            val key = s"Flusher-$this-${rand.nextInt()}"
+            workerSource.sample(WorkerSource.FlushDataTime, key) {
+              if (!task.notifier.hasException) {
+                try {
+                  val flushBeginTime = System.nanoTime()
+                  lastBeginFlushTime.set(index, flushBeginTime)
+                  task.flush()
+                  flushTotalTime.add(System.nanoTime() - flushBeginTime)
+                  flushCount.increment()
+                } catch {
+                  case _: ClosedByInterruptException =>
+                  case e: IOException =>
+                    task.notifier.setException(e)
+                    processIOException(e, DiskStatus.ReadOrWriteFailure)
+                }
+                lastBeginFlushTime.set(index, -1)
+              }
+              returnBuffer(task.buffer)
+              task.notifier.numPendingFlushes.decrementAndGet()
+            }
+          }
+        }
+      }
+      workers(index).setDaemon(true)
+      workers(index).setUncaughtExceptionHandler(new Thread.UncaughtExceptionHandler {
+        override def uncaughtException(t: Thread, e: Throwable): Unit = {
+          logError(s"$this thread terminated.", e)
+        }
+      })
+      workers(index).start()
+    }
+  }
+
+  def getWorkerIndex: Int = synchronized {
+    nextWorkerIndex = (nextWorkerIndex + 1) % threadCount
+    nextWorkerIndex
+  }
+
+  def averageFlushTime(): Long = {
+    logInfo(s"flushCount $flushCount")
+    val currentFlushTime = flushTotalTime.sumThenReset()
+    val currentFlushCount = flushCount.sumThenReset()
+    if (currentFlushCount >= flushAvgTimeMinimumCount) {
+      avgTimeWindow(avgTimeWindowCurrentIndex) = (currentFlushTime, currentFlushCount)
+      avgTimeWindowCurrentIndex = (avgTimeWindowCurrentIndex + 1) % flushAvgTimeWindowSize
+    }
+
+    var totalFlushTime = 0L
+    var totalFlushCount = 0L
+    avgTimeWindow.foreach { case (flushTime, flushCount) =>
+      totalFlushTime = totalFlushTime + flushTime
+      totalFlushCount = totalFlushCount + flushCount
+    }
+
+    if (totalFlushCount != 0) {
+      totalFlushTime / totalFlushCount
+    } else {
+      0L
+    }
+  }
+
+  def takeBuffer(): CompositeByteBuf = {
+    var buffer = bufferQueue.poll()
+    if (buffer == null) {
+      buffer = Unpooled.compositeBuffer(256)
+    }
+    buffer
+  }
+
+  def returnBuffer(buffer: CompositeByteBuf): Unit = {
+    MemoryTracker.instance().releaseDiskBuffer(buffer.readableBytes())
+    buffer.removeComponents(0, buffer.numComponents())
+    buffer.clear()
+
+    bufferQueue.put(buffer)
+  }
+
+  def addTask(task: FlushTask, timeoutMs: Long, workerIndex: Int): Boolean = {
+    workingQueues(workerIndex).offer(task, timeoutMs, TimeUnit.MILLISECONDS)
+  }
+
+  def bufferQueueInfo(): String = s"$this used buffers: ${bufferQueue.size()}"
+
+  def processIOException(e: IOException, deviceErrorType: DiskStatus): Unit
+}
+
+private[worker] class LocalFlusher(
+    workerSource: AbstractSource,
+    val deviceMonitor: DeviceMonitor,
+    threadCount: Int,
+    val mountPoint: String,
+    flushAvgTimeWindowSize: Int,
+    flushAvgTimeMinimumCount: Int,
+    val diskType: StorageInfo.Type) extends Flusher(
+      workerSource,
+      threadCount,
+      flushAvgTimeWindowSize,
+      flushAvgTimeMinimumCount)
+      with DeviceObserver with Logging {
+
+  deviceMonitor.registerFlusher(this)
+
+  override def processIOException(e: IOException, deviceErrorType: DiskStatus): Unit = {
+    stopFlag.set(true)
+    logError(s"$this write failed, report to DeviceMonitor, eception: $e")
+    deviceMonitor.reportDeviceError(mountPoint, e, deviceErrorType)
+  }
+
+  override def notifyError(mountPoint: String,
+                           diskStatus: DiskStatus): Unit = {
+    logError(s"$this is notified Disk $mountPoint $diskStatus! Stop LocalFlusher.")
+    stopFlag.set(true)
+    try {
+      workers.foreach(_.interrupt())
+    } catch {
+      case e: Exception =>
+        logError(s"Exception when interrupt worker: ${workers.mkString(",")}, $e")
+    }
+    workingQueues.foreach { queue =>
+      queue.asScala.foreach { task =>
+        task.buffer.removeComponents(0, task.buffer.numComponents())
+        task.buffer.clear()
+      }
+    }
+    deviceMonitor.unregisterFlusher(this)
+  }
+
+  override def hashCode(): Int = {
+    mountPoint.hashCode()
+  }
+
+  override def equals(obj: Any): Boolean = {
+    obj.isInstanceOf[LocalFlusher] &&
+      obj.asInstanceOf[LocalFlusher].mountPoint.equals(mountPoint)
+  }
+
+  override def toString(): String = {
+    s"LocalFlusher@$flusherId-$mountPoint"
+  }
+}
 
 private[worker] final class StorageManager(
     conf: RssConf,
