--- server-worker/src/main/scala/com/aliyun/emr/rss/service/deploy/worker/storage/DeviceObserver.scala
@@ -1,26 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.aliyun.emr.rss.service.deploy.worker.storage
-
-import com.aliyun.emr.rss.common.meta.DiskStatus
-
-trait DeviceObserver {
-  def notifyError(mountPoint: String, diskStatus: DiskStatus): Unit = {}
-  def notifyHealthy(mountPoint: String): Unit = {}
-  def notifyHighDiskUsage(mountPoint: String): Unit = {}
-}
