--- server-worker/src/main/scala/com/aliyun/emr/rss/service/deploy/worker/storage/local/LocalFlushTask.scala
@@ -1,34 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.aliyun.emr.rss.service.deploy.worker.storage.local
-
-import java.nio.channels.FileChannel
-
-import io.netty.buffer.CompositeByteBuf
-
-import com.aliyun.emr.rss.service.deploy.worker.FileWriter
-import com.aliyun.emr.rss.service.deploy.worker.storage.FlushTask
-
-private[worker] class LocalFlushTask(
-    buffer: CompositeByteBuf,
-    fileChannel: FileChannel,
-    notifier: FileWriter.FlushNotifier) extends FlushTask(buffer, notifier) {
-  override def flush(): Unit = {
-    fileChannel.write(buffer.nioBuffers())
-  }
-}
